backend/embedding_service/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pydantic==2.9.2
